requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
--no-binary psycopg2

arrow==1.1.0              # via django-q
blessed==1.18.0           # via django-q
certifi==2020.6.20        # via requests, skyfield
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
croniter==1.0.13
cryptography==3.1.1       # via social-auth-core
defusedxml==0.6.0         # via python3-openid, social-auth-core
django==2.2.13
git+https://github.com/andersinno/django-activity-stream.git#egg=django-activity-stream==0.7.0
django-admin-sortable2==1.0.2
django-allauth==0.42.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-extended-choices==1.3.3
django-filter==2.3.0
django-helusers==0.5.6
django-jsonfield-compat==0.4.4
django-picklefield==3.0.1  # via django-q
django-private-storage==2.2.2
django-q==1.3.6
djangorestframework==3.10.0
djangorestframework-gis==0.15
docxtpl==0.10.0
drf-extensions==0.4.0
drf-oidc-auth==0.10.0     # via django-helusers
ecdsa==0.14.1             # via python-jose
et-xmlfile==1.0.1         # via openpyxl
future==0.18.2            # via pyjwkest
idna==2.9                 # via requests
jdcal==1.4.1              # via openpyxl
jinja2==2.11.2            # via docxtpl
jplephem==2.15            # via skyfield
lunardate==0.2.0          # via workalendar
lxml==4.5.1               # via docxtpl, python-docx
markupsafe==1.1.1         # via jinja2
numpy==1.19.2             # via jplephem, skyfield
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
openpyxl==3.0.3
pillow==8.3.1
psycopg2==2.8.5
pyasn1==0.4.8             # via python-jose, rsa
pycalverter==1.6.1        # via workalendar
pycparser==2.20           # via cffi
pycryptodomex==3.9.8      # via pyjwkest
pyjwkest==1.4.2           # via drf-oidc-auth
pyjwt==1.7.1              # via social-auth-core
pyluach==1.1.1            # via workalendar
python-dateutil==2.8.1    # via arrow, croniter, workalendar
python-docx==0.8.10       # via docxtpl
python-jose==3.2.0        # via django-helusers
python3-openid==3.2.0     # via django-allauth, social-auth-core
pytz==2020.1              # via django, workalendar
raven==6.10.0
requests==2.24.0          # via django-allauth, django-helusers, pyjwkest, requests-oauthlib, social-auth-core
requests-oauthlib==1.3.0  # via django-allauth, social-auth-core
rsa==4.6                  # via python-jose
sgp4==2.13                # via skyfield
six==1.15.0               # via blessed, cryptography, django-extended-choices, docxtpl, ecdsa, pyjwkest, python-dateutil, python-jose, social-auth-app-django, social-auth-core
skyfield==1.30            # via workalendar
skyfield-data==1.1.0      # via workalendar
social-auth-app-django==4.0.0
social-auth-core==3.3.3   # via social-auth-app-django
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
wcwidth==0.2.5            # via blessed
workalendar==11.0.1

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via workalendar
